Reorganize trie/ floder structure: preparation for cursor implementation
