+++ core/blockchain/impl/types.cpp
@@ -7,8 +7,8 @@
 #include "blockchain/impl/storage_util.hpp"
 #include "common/visitor.hpp"
 #include "storage/in_memory/in_memory_storage.hpp"
-#include "storage/trie/impl/trie_serializer_impl.hpp"
-#include "storage/trie/impl/polkadot_trie_impl.hpp"
+#include "storage/trie/polkadot_trie/polkadot_trie_impl.hpp"
+#include "storage/trie/serialization/trie_serializer_impl.hpp"
 
 OUTCOME_CPP_DEFINE_CATEGORY(kagome::blockchain, Error, e) {
   switch (e) {
