+++ core/storage/trie/CMakeLists.txt
@@ -1,4 +1,6 @@
 # Copyright Soramitsu Co., Ltd. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
+add_subdirectory(polkadot_trie)
+add_subdirectory(serialization)
 add_subdirectory(impl)
