+++ core/storage/trie/polkadot_trie/CMakeLists.txt
@@ -1,0 +1,34 @@
+# Copyright Soramitsu Co., Ltd. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+add_library(polkadot_node
+    polkadot_node.cpp
+    )
+target_link_libraries(polkadot_node
+    buffer
+    )
+kagome_install(polkadot_node)
+
+add_library(trie_error
+    trie_error.cpp
+    )
+target_link_libraries(trie_error
+    outcome
+    )
+kagome_install(trie_error)
+
+add_library(polkadot_trie
+    polkadot_trie_impl.cpp
+    )
+target_link_libraries(polkadot_trie
+    polkadot_node
+    trie_error
+    )
+kagome_install(polkadot_trie)
+
+add_library(polkadot_trie_factory
+    polkadot_trie_factory_impl.cpp
+    )
+target_link_libraries(polkadot_trie_factory
+    polkadot_trie
+    )
