+++ core/storage/trie/polkadot_trie/polkadot_trie_impl.cpp
@@ -3,12 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "storage/trie/impl/polkadot_trie_impl.hpp"
+#include "storage/trie/polkadot_trie/polkadot_trie_impl.hpp"
 
 #include <functional>
 #include <utility>
 
-#include "storage/trie/impl/trie_error.hpp"
+#include "storage/trie/polkadot_trie/trie_error.hpp"
 
 using kagome::common::Buffer;
