+++ core/storage/trie/serialization/ordered_trie_hash.hpp
@@ -9,8 +9,8 @@
 #include "common/buffer.hpp"
 #include "crypto/hasher.hpp"
 #include "scale/scale.hpp"
-#include "storage/trie/impl/polkadot_codec.hpp"
-#include "storage/trie/impl/polkadot_trie_impl.hpp"
+#include "storage/trie/polkadot_trie/polkadot_trie_impl.hpp"
+#include "storage/trie/serialization/polkadot_codec.hpp"
 
 namespace kagome::storage::trie {
