+++ core/storage/trie/serialization/polkadot_codec.hpp
@@ -10,9 +10,9 @@
 #include <optional>
 #include <string>
 
+#include "storage/trie/serialization/buffer_stream.hpp"
 #include "storage/trie/codec.hpp"
-#include "storage/trie/impl/buffer_stream.hpp"
-#include "storage/trie/impl/polkadot_node.hpp"
+#include "storage/trie/polkadot_trie/polkadot_node.hpp"
 
 namespace kagome::storage::trie {
