+++ test/core/runtime/wasm_executor_test.cpp
@@ -16,11 +16,11 @@
 #include "mock/core/storage/changes_trie/changes_tracker_mock.hpp"
 #include "runtime/binaryen/runtime_manager.hpp"
 #include "storage/in_memory/in_memory_storage.hpp"
-#include "storage/trie/impl/polkadot_codec.hpp"
-#include "storage/trie/impl/polkadot_trie_factory_impl.hpp"
-#include "storage/trie/impl/trie_serializer_impl.hpp"
 #include "storage/trie/impl/trie_storage_backend_impl.hpp"
 #include "storage/trie/impl/trie_storage_impl.hpp"
+#include "storage/trie/polkadot_trie/polkadot_trie_factory_impl.hpp"
+#include "storage/trie/serialization/polkadot_codec.hpp"
+#include "storage/trie/serialization/trie_serializer_impl.hpp"
 #include "testutil/outcome.hpp"
 #include "testutil/runtime/common/basic_wasm_provider.hpp"
