+++ test/core/storage/trie/trie_storage/polkadot_codec_nibbles_test.cpp
@@ -6,7 +6,7 @@
 #include <memory>
 
 #include <gtest/gtest.h>
-#include "storage/trie/impl/polkadot_codec.hpp"
+#include "storage/trie/serialization/polkadot_codec.hpp"
 
 using namespace kagome;
 using namespace common;
