+++ test/core/storage/trie/trie_storage/trie_batch_test.cpp
@@ -8,11 +8,11 @@
 
 #include "storage/changes_trie/impl/storage_changes_tracker_impl.hpp"
 #include "storage/in_memory/in_memory_storage.hpp"
-#include "storage/trie/impl/polkadot_trie_factory_impl.hpp"
-#include "storage/trie/impl/trie_error.hpp"
-#include "storage/trie/impl/trie_serializer_impl.hpp"
 #include "storage/trie/impl/trie_storage_backend_impl.hpp"
 #include "storage/trie/impl/trie_storage_impl.hpp"
+#include "storage/trie/polkadot_trie/polkadot_trie_factory_impl.hpp"
+#include "storage/trie/polkadot_trie/trie_error.hpp"
+#include "storage/trie/serialization/trie_serializer_impl.hpp"
 #include "storage/trie/trie_batches.hpp"
 #include "testutil/literals.hpp"
 #include "testutil/outcome.hpp"
